backend/signup/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cloudinary==1.31.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
